requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
Pillow==10.4.0
aiohttp==3.9.3
